requirements.txt
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.0.0
tabulate>=0.9.0
platformdirs>=4.0.0
PyYAML>=6.0
pytest>=7.4.0
